editor: refactor typewriter, remove fonts
refactor typewriter: knotCaret and typewriter is two seperated features,
but typewriter scroll function is depends on caret's absolute position.
in the future it has to be togglable. So had to divide the css file to
fit in it's scope.

remove fonts: now on tailwind and ui library is responsible to load
fonts, remove downloaded font files and remove it from editor.css.