feat: use nodesBetween function to loop over nodes and grab headings
`nodesBetween` function loops over entire document and checks if the
node type is heading. If it is, it stores in a list and set the last
heading as a variable.

It also handles user selection(focus on node) by limiting the range to be
from the start of the document to where the caret is located at.